🔨refactor: Rename eventhandler -> eventsourced
