+++ src/main/resources/reference.conf
@@ -88,7 +88,7 @@ lerna.akka.entityreplication {
     }
   }
 
-  raft.eventhandler {
+  raft.eventsourced {
     // Settings for saving committed events from each RaftActor
     commit-log-store {
       // Retry setting to prevent events from being lost if commit-log-store(sharding) stops temporarily
