+++ src/main/scala/lerna/akka/entityreplication/raft/eventsourced/CommitLogStoreActor.scala
@@ -1,4 +1,4 @@
-package lerna.akka.entityreplication.raft.eventhandler
+package lerna.akka.entityreplication.raft.eventsourced
 
 import akka.Done
 import akka.actor.{ ActorRef, ActorSystem, Props }
@@ -43,7 +43,7 @@ class CommitLogStoreActor(typeName: String) extends PersistentActor {
   // TODO: 複数 Raft(typeName) に対応するために typeName ごとに cassandra-journal.keyspace を分ける
   override def journalPluginId: String =
     context.system.settings.config
-      .getString("lerna.akka.entityreplication.raft.eventhandler.persistence.journal.plugin")
+      .getString("lerna.akka.entityreplication.raft.eventsourced.persistence.journal.plugin")
 
   // TODO: Use snapshot for efficient recovery after reboot
   override def snapshotPluginId: String = "akka.persistence.no-snapshot-store"
