+++ src/main/scala/lerna/akka/entityreplication/raft/eventsourced/ShardedCommitLogStore.scala
@@ -1,9 +1,9 @@
-package lerna.akka.entityreplication.raft.eventhandler
+package lerna.akka.entityreplication.raft.eventsourced
 
 import akka.actor.{ ActorSystem, Scheduler }
 import akka.pattern.ask
 import akka.util.Timeout
-import lerna.akka.entityreplication.raft.eventhandler.CommitLogStore.ReplicationId
+import lerna.akka.entityreplication.raft.eventsourced.CommitLogStore.ReplicationId
 import lerna.akka.entityreplication.raft.model.LogEntryIndex
 
 import scala.jdk.DurationConverters._
@@ -14,7 +14,7 @@ class ShardedCommitLogStore(typeName: String, system: ActorSystem) extends Commi
   private implicit val scheduler: Scheduler = system.scheduler
 
   private val config =
-    system.settings.config.getConfig("lerna.akka.entityreplication.raft.eventhandler.commit-log-store")
+    system.settings.config.getConfig("lerna.akka.entityreplication.raft.eventsourced.commit-log-store")
   private val retryAttempts                  = config.getInt("retry.attempts")
   private val retryDelay: FiniteDuration     = config.getDuration("retry.delay").toScala
   private implicit val retryTimeout: Timeout = Timeout(retryDelay)
