+++ src/multi-jvm/resources/multi-jvm-testing.conf
@@ -34,7 +34,7 @@ lerna.akka.entityreplication.raft {
   }
 }
 
-lerna.akka.entityreplication.raft.eventhandler {
+lerna.akka.entityreplication.raft.eventsourced {
   persistence {
     journal.plugin = "akka.persistence.journal.proxy"
   }
