+++ src/test/resources/akka-entity-replication-with-cassandra.conf
@@ -3,8 +3,8 @@ lerna.akka.entityreplication.raft.persistence {
   snapshot-store.plugin = "akka-entity-replication.raft.persistence.cassandra.snapshot-store"
 }
 
-lerna.akka.entityreplication.raft.eventhandler.persistence {
-  journal.plugin = "akka-entity-replication.eventhandler.persistence.cassandra.journal"
+lerna.akka.entityreplication.raft.eventsourced.persistence {
+  journal.plugin = "akka-entity-replication.eventsourced.persistence.cassandra.journal"
 }
 
 // The settings for Cassandra persistence plugin to handle command side
@@ -53,7 +53,7 @@ akka-entity-replication.raft.persistence.cassandra = ${akka.persistence.cassandr
 
 // The settings for Cassandra persistence plugin to handle query side
 // You can set anything name for the root key of these settings
-akka-entity-replication.eventhandler.persistence.cassandra = ${akka.persistence.cassandra} {
+akka-entity-replication.eventsourced.persistence.cassandra = ${akka.persistence.cassandra} {
   // Profile to use.
   // See https://docs.datastax.com/en/developer/java-driver/latest/manual/core/configuration/ for overriding any settings
   read-profile = "akka-entity-replication-profile"
