+++ src/test/resources/application.conf
@@ -32,7 +32,7 @@ lerna.akka.entityreplication.raft {
   }
 }
 
-lerna.akka.entityreplication.raft.eventhandler {
+lerna.akka.entityreplication.raft.eventsourced {
   persistence {
     journal.plugin = "inmemory-journal"
   }
